feat: 스터디 상세 정보 비즈니스 로직에서 스터디 비즈니스 로직으로 변경
