+++ src/main/java/com/gdschongik/gdsc/domain/study/api/MentorStudyController.java
@@ -1,6 +1,7 @@
 package com.gdschongik.gdsc.domain.study.api;
 
 import com.gdschongik.gdsc.domain.study.application.MentorStudyService;
+import com.gdschongik.gdsc.domain.study.dto.request.StudyDetailUpdateRequest;
 import com.gdschongik.gdsc.domain.study.dto.response.MentorStudyResponse;
 import com.gdschongik.gdsc.domain.study.dto.response.StudyStudentResponse;
 import io.swagger.v3.oas.annotations.Operation;
@@ -8,10 +9,7 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @Tag(name = "Mentor Study", description = "멘토 스터디 API입니다.")
 @RestController
@@ -20,6 +18,14 @@ import org.springframework.web.bind.annotation.RestController;
 public class MentorStudyController {
 
     private final MentorStudyService mentorStudyService;
+
+    @Operation(summary = "스터디 상세 정보 작성", description = "스터디 상세 정보를 작성합니다.")
+    @PatchMapping("/{studyId}")
+    public ResponseEntity<Void> updateStudyDetail(
+            @PathVariable Long studyId, @RequestBody StudyDetailUpdateRequest request) {
+        mentorStudyService.updateStudyDetail(studyId, request);
+        return ResponseEntity.ok().build();
+    }
 
     @Operation(summary = "내 스터디 조회", description = "내가 멘토로 있는 스터디를 조회합니다.")
     @GetMapping("/me")
