+++ src/main/java/com/gdschongik/gdsc/domain/study/api/MentorStudyDetailController.java
@@ -2,7 +2,6 @@ package com.gdschongik.gdsc.domain.study.api;
 
 import com.gdschongik.gdsc.domain.study.application.MentorStudyDetailService;
 import com.gdschongik.gdsc.domain.study.dto.request.AssignmentCreateUpdateRequest;
-import com.gdschongik.gdsc.domain.study.dto.request.StudyDetailUpdateRequest;
 import com.gdschongik.gdsc.domain.study.dto.response.AssignmentResponse;
 import com.gdschongik.gdsc.domain.study.dto.response.StudySessionResponse;
 import io.swagger.v3.oas.annotations.Operation;
@@ -27,14 +26,6 @@ import org.springframework.web.bind.annotation.RestController;
 public class MentorStudyDetailController {
 
     private final MentorStudyDetailService mentorStudyDetailService;
-
-    @Operation(summary = "스터디 상세 정보 작성", description = "스터디 상세 정보를 작성합니다.")
-    @PatchMapping
-    public ResponseEntity<Void> updateStudyDetail(
-            @RequestParam(name = "studyId") Long studyId, @RequestBody StudyDetailUpdateRequest request) {
-        mentorStudyDetailService.updateStudyDetail(studyId, request);
-        return ResponseEntity.ok().build();
-    }
 
     @Operation(summary = "스터디 과제 수정", description = "멘토만 과제를 수정할 수 있습니다.")
     @PatchMapping("/{studyDetailId}/assignments")
