+++ src/main/java/com/gdschongik/gdsc/domain/study/application/MentorStudyDetailService.java
@@ -4,23 +4,14 @@ import static com.gdschongik.gdsc.global.exception.ErrorCode.*;
 
 import com.gdschongik.gdsc.domain.member.domain.Member;
 import com.gdschongik.gdsc.domain.study.dao.StudyDetailRepository;
-import com.gdschongik.gdsc.domain.study.dao.StudyRepository;
-import com.gdschongik.gdsc.domain.study.domain.Study;
 import com.gdschongik.gdsc.domain.study.domain.StudyDetail;
 import com.gdschongik.gdsc.domain.study.domain.StudyDetailValidator;
-import com.gdschongik.gdsc.domain.study.domain.StudyValidator;
 import com.gdschongik.gdsc.domain.study.dto.request.AssignmentCreateUpdateRequest;
-import com.gdschongik.gdsc.domain.study.dto.request.StudyDetailUpdateRequest;
-import com.gdschongik.gdsc.domain.study.dto.request.StudySessionCreateRequest;
 import com.gdschongik.gdsc.domain.study.dto.response.AssignmentResponse;
 import com.gdschongik.gdsc.domain.study.dto.response.StudySessionResponse;
 import com.gdschongik.gdsc.global.exception.CustomException;
 import com.gdschongik.gdsc.global.util.MemberUtil;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -34,8 +25,6 @@ public class MentorStudyDetailService {
     private final MemberUtil memberUtil;
     private final StudyDetailRepository studyDetailRepository;
     private final StudyDetailValidator studyDetailValidator;
-    private final StudyRepository studyRepository;
-    private final StudyValidator studyValidator;
 
     @Transactional(readOnly = true)
     public List<AssignmentResponse> getWeeklyAssignments(Long studyId) {
@@ -100,40 +89,5 @@ public class MentorStudyDetailService {
     public List<StudySessionResponse> getSessions(Long studyId) {
         List<StudyDetail> studyDetails = studyDetailRepository.findAllByStudyId(studyId);
         return studyDetails.stream().map(StudySessionResponse::from).toList();
-    }
-
-    // TODO session -> curriculum 변경
-    @Transactional
-    public void updateStudyDetail(Long studyId, StudyDetailUpdateRequest request) {
-        Member currentMember = memberUtil.getCurrentMember();
-        Study study = studyRepository.findById(studyId).orElseThrow(() -> new CustomException(STUDY_NOT_FOUND));
-        studyValidator.validateStudyMentor(currentMember, study);
-
-        List<StudyDetail> studyDetails = studyDetailRepository.findAllByStudyId(studyId);
-        studyDetailValidator.validateUpdateStudyDetail(studyDetails, request.studySessions());
-
-        study.update(request.notionLink(), request.introduction());
-        studyRepository.save(study);
-        log.info("[MentorStudyDetailService] 스터디 기본 정보 수정 완료: studyId={}", studyId);
-
-        Map<Long, StudySessionCreateRequest> requestMap = request.studySessions().stream()
-                .collect(Collectors.toMap(StudySessionCreateRequest::studyDetailId, Function.identity()));
-
-        List<StudyDetail> updatedStudyDetails = new ArrayList<>();
-        for (StudyDetail studyDetail : studyDetails) {
-            Long id = studyDetail.getId();
-            StudySessionCreateRequest matchingSession = requestMap.get(id);
-
-            studyDetail.updateSession(
-                    studyDetail.getStudy().getPeriod().getStartDate(),
-                    matchingSession.title(),
-                    matchingSession.description(),
-                    matchingSession.difficulty(),
-                    matchingSession.status());
-
-            updatedStudyDetails.add(studyDetail);
-        }
-        studyDetailRepository.saveAll(updatedStudyDetails);
-        log.info("[MentorStudyDetailService] 스터디 상세정보 커리큘럼 작성 완료: studyDetailId={}", studyDetails);
     }
 }
